API reference

Raven.constructorof — Function

constructorof(T::Type) -> constructor

Return an object constructor that can be used to construct objects of type T from their field values. Typically, constructor will be the type T with all parameters removed:

julia> struct T{A,B}
           a::A
           b::B
       end

julia> Raven.constructorof(T{Int,Int})
T

The returned constructor is used to unflatten objects hierarchical types from a list of their values. For example, in this case T(1,2) constructs an object where T.a==1 an T.b==2.

The method constructorof should be defined for types that are not constructed from a tuple of their values.

Raven.flatten — Method

flatten(obj, use=Real)

Flattens a hierarchical type to a tuple with elements of type use.

Examples

julia> flatten((a=(Complex(1, 2), 3), b=4))
(1, 2, 3, 4)

To convert the tuple to a vector, simply use [flatten(x)...], or using static arrays to avoid allocations: SVector(flatten(x)).

Raven.get_backend — Method

get_backend(::Type{T}) -> Type

Returns the KernelAbstractions backend to use with kernels where A is an argument.

Raven.materializequadranttoglobalid — Method

materializequadranttoglobalid(forest, ghost)

Generate the global ids for quadrants in the forest and the ghost layer.

Raven.numbercontiguous — Method

numbercontiguous(T, A; by = identity)

Renumbers A contiguously in an Array{T} and returns it. The function by is a mapping for the elements of A used during element comparison, similar to sort.

Examples

julia> Raven.numbercontiguous(Int32, [13, 4, 5, 1, 5])
5-element Vector{Int32}:
 4
 2
 3
 1
 3

julia> Raven.numbercontiguous(Int32, [13, 4, 5, 1, 5]; by = x->-x)
5-element Vector{Int32}:
 1
 3
 2
 4
 2

Raven.parenttype — Method

parent_type(::Type{T}) -> Type

Returns the parent array that type T wraps.

Raven.pin — Method

pin(T::Type, A::Array)

Pins the host array A for copying to arrays of type T

Raven.unflatten — Function

unflatten(T::Type, data, use::Type=Real)

Construct an object from Tuple or Vector data and a Type T. The data should be at least as long as the queried fields (of type use) in T.

Examples

julia> unflatten(Tuple{Tuple{Int,Int},Complex{Int,Int}}, (1, 2, 3, 4))
((1, 2), 3 + 4im)