API reference

Raven.get_backend — Method

get_backend(::Type{T}) -> Type

Returns the KernelAbstractions backend to use with kernels where A is an argument.

Raven.materializequadranttoglobalid — Method

materializequadranttoglobalid(forest, ghost)

Generate the global ids for quadrants in the forest and the ghost layer.

Raven.numbercontiguous — Method

numbercontiguous(T, A; by = identity)

Renumbers A contiguously in an Array{T} and returns it. The function by is a mapping for the elements of A used during element comparison, similar to sort.

Examples

julia> Raven.numbercontiguous(Int32, [13, 4, 5, 1, 5])
5-element Vector{Int32}:
 4
 2
 3
 1
 3

julia> Raven.numbercontiguous(Int32, [13, 4, 5, 1, 5]; by = x->-x)
5-element Vector{Int32}:
 1
 3
 2
 4
 2

Raven.parenttype — Method

parent_type(::Type{T}) -> Type

Returns the parent array that type T wraps.

Raven.pin — Method

pin(T::Type, A::Array)

Pins the host array A for copying to arrays of type T